Refactor MonitorBeanDefinitionTag
